fixed timestamp to handle non-integer frame rates

+++ libftl/ftl_private.h
@@ -133,8 +133,8 @@ typedef struct {
 typedef struct {
 	uint8_t payload_type;
 	uint32_t ssrc;
-	uint32_t timestamp;
-	uint32_t timestamp_step;
+	float timestamp;
+	float timestamp_step;
 	uint16_t seq_num;
 	int64_t min_nack_rtt;
 	int64_t max_nack_rtt;
